Generateed SpawnSystem
Switched spawning tasks from GameManager to a new SpawnSystem prefab
