+++ Assets/Scripts/Spawn System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f8039ec84939194888644ad2f7f7869
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
