+++ Assets/Scripts/Spawn System/SpawnSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6dce32e7ffad20c4ebaf61a526b853ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
