+++ Assets/prefabs/Level/SpawnSystem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bc3ae821301b64e4cb2bb65b9e9a71ad
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
